Fix own message read marker setting
This will always correctly mark rooms as read unread.

This still has a minimal corner case. But I think it is tolerable,
whereas a fix would be very hard (and should be done in the library):

When we get in a sync an own event and after that non-local events
which we have not read, the read-marker will stay before our own
messages. This is the case, since we do not know, if we read the
messages after our own at this point. If we would set the readmarker to
our own message in this moment it could lead to us wrongly marking the
following messages as unread.